app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
mangum==0.17.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.29.0
